refactor(ipc): 重构IPC架构并更新文档
## 主要变更

### 架构重构
- 引入Channel抽象层统一读写通道管理 (+1049行)
- 新增ChannelRegistry支持多通道复用 (+219行)
- 实现IPCFactory工厂模式简化创建流程 (+152行)
- 优化Publisher/Subscriber API，移除泛型模板

### 核心功能增强
- Publisher支持Lambda写入、缓冲区拷贝、定向发送
- Subscriber支持批量接收(Vector<Sample>)、Lambda读取
- 新增Connect/Disconnect连接管理
- 实现STmin动态更新(统一微秒单位)
- 完善引用计数和共享内存生命周期管理

### 测试用例扩展
- 新增SPSC/MPSC/MPMC camera_fusion示例 (+2166行)
- 新增property_board服务属性管理示例 (+438行)
- 新增ipc_service完整服务示例 (+389行)
- 移除按模式分离的冗余测试(-1584行)
- 新增IPCFactory单元测试 (+117行)

### 文档更新
- 更新IPC_DESIGN_ARCHITECTURE.md架构设计
  - 修正API签名(移除泛型模板)
  - 添加批量接收、Lambda方式等新API
  - 更新创建示例匹配实际代码
  - 近期实现更新: shm生命周期/订阅者退出/STmin单位
- 更新INDEX.md和README文档索引(2026-02-01)
- 删除过时临时分析文档(-4621行)

### 代码优化
- 重构ControlBlock结构提升性能
- 修复header guard不一致问题(CRawAtomic.hpp)
- 统一错误码和类型定义
- 优化CMakeLists.txt测试配置

## 统计
- 新增文件: 12个 (+6978行)
- 删除文件: 26个 (-8297行)
- 修改文件: 27个
- 净变化: -1319行(代码更精简)

+++ source/inc/ipc/ChunkPoolAllocator.hpp
@@ -13,8 +13,8 @@
 #include "IPCTypes.hpp"
 #include "ChunkHeader.hpp"
 #include "ControlBlock.hpp"
+#include "SharedMemoryManager.hpp"
 #include "CResult.hpp"
-#include "COptional.hpp"
 
 namespace lap
 {
@@ -65,27 +65,27 @@ namespace ipc
          * @brief Allocate a chunk (lock-free)
          * @return Optional chunk index (empty if pool exhausted)
          */
-        UInt32 Allocate() noexcept;
+        UInt16 Allocate() noexcept;
         
         /**
          * @brief Deallocate a chunk (lock-free)
          * @param chunk_index Chunk index to free
          */
-        void Deallocate(UInt32 chunk_index) noexcept;
+        void Deallocate( UInt16 chunk_index ) noexcept;
         
         /**
          * @brief Get chunk header by index
          * @param chunk_index Chunk index
          * @return Chunk header pointer
          */
-        ChunkHeader* GetChunkHeader(UInt32 chunk_index) const noexcept;
+        ChunkHeader* GetChunkHeader( UInt16 chunk_index ) const noexcept;
         
         /**
          * @brief Get chunk payload by index
          * @param chunk_index Chunk index
          * @return Payload pointer
          */
-        void* GetChunkPayload(UInt32 chunk_index) const noexcept;
+        void* GetChunkPayload( UInt16 chunk_index ) const noexcept;
         
         /**
          * @brief Check if pool is exhausted
@@ -100,7 +100,7 @@ namespace ipc
          * @brief Get allocated chunk count
          * @return Number of allocated chunks
          */
-        inline UInt32 GetAllocatedCount() const noexcept
+        inline UInt16 GetAllocatedCount() const noexcept
         {
             return control_->header.max_chunks - control_->pool_state.remain_count.load(std::memory_order_acquire);
         }
@@ -109,7 +109,7 @@ namespace ipc
          * @brief Get maximum chunks
          * @return Maximum number of chunks
          */
-        inline UInt32 GetMaxChunks() const noexcept
+        inline UInt16 GetMaxChunks() const noexcept
         {
             return control_->header.max_chunks;
         }
@@ -120,7 +120,7 @@ namespace ipc
          * @param index Chunk index
          * @return Chunk header address
          */
-        inline ChunkHeader* GetChunkAt(UInt32 index) const noexcept;
+        inline ChunkHeader* GetChunkAt(UInt16 index) const noexcept;
         
         void* base_addr_;              ///< Shared memory base address
         ControlBlock* control_;        ///< Control block
